requirements.txt
aiogram>=3.27,<4
python-dotenv>=1.1,<2
openpyxl>=3.1,<4
tzdata>=2025.2
